move TripFormContainer to separate file
